requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
feedparser==6.0.10
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
pyfiglet==1.0.2
Pygments==2.17.2
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
sgmllib3k==1.0.0
Unidecode==1.3.8
urllib3==2.2.1
